requirements.txt
Package                   Version
------------------------- --------------------
aiofiles                  24.1.0
aiohappyeyeballs          2.4.4
aiohttp                   3.10.11
aiosignal                 1.3.1
annotated-types           0.7.0
anyio                     4.5.2
apturl                    0.5.2
argon2-cffi               23.1.0
argon2-cffi-bindings      21.2.0
arrow                     1.3.0
asttokens                 3.0.0
async-lru                 2.0.4
async-timeout             4.0.3
attrs                     25.1.0
babel                     2.17.0
backcall                  0.2.0
bcrypt                    3.1.7
beautifulsoup4            4.13.3
black                     24.8.0
bleach                    6.1.0
blinker                   1.8.2
Brlapi                    0.7.0
cassandra-pylib           0.0.0
certifi                   2025.1.31
cffi                      1.17.1
chardet                   3.0.4
charset-normalizer        3.4.1
ci-info                   0.3.0
click                     8.1.8
colorama                  0.4.3
comm                      0.2.2
command-not-found         0.3
configobj                 5.0.9
configparser              7.1.0
cryptography              2.8
cupshelpers               1.0
dataclasses-json          0.6.7
dbus-python               1.2.16
debugpy                   1.8.12
decorator                 5.2.1
defer                     1.0.6
defusedxml                0.7.1
distro                    1.9.0
distro-info               0.23+ubuntu1.1
docx2txt                  0.8
duplicity                 0.8.12.0
entrypoints               0.3
et_xmlfile                2.0.0
etelemetry                0.3.1
exceptiongroup            1.2.2
executing                 2.2.0
fasteners                 0.14.1
fastjsonschema            2.21.1
filelock                  3.16.1
fitz                      0.0.1.dev2
Flask                     3.0.3
Flask-WTF                 1.2.1
fqdn                      1.5.1
frontend                  0.0.3
frozenlist                1.5.0
future                    0.18.2
greenlet                  3.1.1
h11                       0.14.0
httpcore                  1.0.7
httplib2                  0.14.0
httpx                     0.28.1
idna                      3.10
importlib_metadata        8.5.0
importlib_resources       6.4.5
ipykernel                 6.29.5
ipython                   8.12.3
ipython_genutils          0.2.0
ipywidgets                8.1.5
isodate                   0.6.1
isoduration               20.11.0
itsdangerous              2.2.0
jedi                      0.19.2
Jinja2                    3.1.5
jiter                     0.8.2
joblib                    1.4.2
json5                     0.10.0
jsonpatch                 1.33
jsonpointer               3.0.0
jsonschema                4.23.0
jsonschema-specifications 2023.12.1
jupyter                   1.1.1
jupyter_client            8.6.3
jupyter-console           6.6.3
jupyter_core              5.7.2
jupyter-events            0.10.0
jupyter-lsp               2.2.5
jupyter_server            2.14.2
jupyter_server_terminals  0.5.3
jupyterlab                4.3.5
jupyterlab_pygments       0.3.0
jupyterlab_server         2.27.3
jupyterlab_widgets        3.0.13
kazam                     1.4.5
keyring                   18.0.1
langchain                 0.2.17
langchain-community       0.2.19
langchain-core            0.2.43
langchain-ollama          0.1.3
langchain-text-splitters  0.2.4
langsmith                 0.1.147
language-selector         0.1
launchpadlib              1.10.13
lazr.restfulclient        0.14.2
lazr.uri                  1.0.3
libvirt-python            6.1.0
lockfile                  0.12.2
looseversion              1.3.0
louis                     3.12.0
lxml                      5.3.1
macaroonbakery            1.3.1
Mako                      1.1.0
MarkupSafe                2.1.5
marshmallow               3.22.0
matplotlib-inline         0.1.7
mistune                   3.1.2
monotonic                 1.5
multidict                 6.1.0
mypy-extensions           1.0.0
nbclient                  0.10.1
nbconvert                 7.16.6
nbformat                  5.10.4
nest-asyncio              1.6.0
netifaces                 0.10.4
networkx                  3.1
nibabel                   5.2.1
nipype                    1.8.6
nltk                      3.9.1
notebook                  7.3.2
notebook_shim             0.2.4
numpy                     1.24.4
oauthlib                  3.1.0
ollama                    0.4.7
openai                    1.64.0
openpyxl                  3.1.5
orjson                    3.10.15
overrides                 7.7.0
packaging                 24.2
pandas                    2.0.3
pandocfilters             1.5.1
paramiko                  2.6.0
parso                     0.8.4
pathlib                   1.0.1
pathspec                  0.12.1
pdf2image                 1.17.0
pexpect                   4.9.0
pickleshare               0.7.5
pillow                    10.4.0
pip                       25.0.1
pkgutil_resolve_name      1.3.10
platformdirs              4.3.6
prometheus_client         0.21.1
prompt_toolkit            3.0.50
propcache                 0.2.0
protobuf                  3.6.1
prov                      2.0.1
psutil                    7.0.0
ptyprocess                0.7.0
pure_eval                 0.2.3
pycairo                   1.16.2
pycparser                 2.22
pycups                    1.9.73
pydantic                  2.10.6
pydantic_core             2.27.2
pydot                     3.0.4
Pygments                  2.19.1
PyGObject                 3.36.0
PyJWT                     1.7.1
pymacaroons               0.13.0
PyMuPDF                   1.24.11
PyNaCl                    1.3.0
pyparsing                 3.1.4
pyRFC3339                 1.1
pytesseract               0.3.13
python-apt                2.0.1+ubuntu0.20.4.1
python-dateutil           2.9.0.post0
python-debian             0.1.36+ubuntu1.1
python-dotenv             1.0.1
python-json-logger        3.2.1
pytz                      2025.1
pyxdg                     0.26
pyxnat                    1.6.3
PyYAML                    6.0.2
pyzmq                     26.2.1
rdflib                    6.3.2
referencing               0.35.1
regex                     2024.11.6
requests                  2.32.3
requests-toolbelt         1.0.0
requests-unixsocket       0.2.0
rfc3339-validator         0.1.4
rfc3986-validator         0.1.1
rpds-py                   0.20.1
scipy                     1.10.1
SecretStorage             2.3.1
Send2Trash                1.8.3
setuptools                75.3.0
simplejson                3.16.0
six                       1.17.0
sniffio                   1.3.1
soupsieve                 2.6
SQLAlchemy                2.0.38
stack-data                0.6.3
starlette                 0.44.0
systemd-python            234
tenacity                  8.5.0
terminado                 0.18.1
tesseract                 0.1.3
tinycss2                  1.2.1
tomli                     2.2.1
tornado                   6.4.2
tqdm                      4.67.1
traitlets                 5.14.3
traits                    6.3.2
types-python-dateutil     2.9.0.20241206
typing_extensions         4.12.2
typing-inspect            0.9.0
tzdata                    2025.1
ubuntu-drivers-common     0.0.0
ubuntu-pro-client         8001
ufw                       0.36
unattended-upgrades       0.1
uri-template              1.3.0
urllib3                   2.2.3
usb-creator               0.3.7
uvicorn                   0.33.0
wadllib                   1.3.3
wcwidth                   0.2.13
webcolors                 24.8.0
webencodings              0.5.1
websocket-client          1.8.0
Werkzeug                  3.0.6
wheel                     0.45.1
widgetsnbextension        4.0.13
WTForms                   3.1.2
xkit                      0.0.0
yarl                      1.15.2
zipp                      3.20.2